Major SConstruct upgrade; Massive VC9 /Wall warning purge; Some minor corrections; Cleaner batch file upgrades;

+++ include/Zeni/Core.h
@@ -102,13 +102,13 @@ namespace Zeni {
     static bool delete_file(const std::string &file_path); ///< Delete a file
     static bool copy_file(const std::string &from, const std::string &to); ///< Copy a file from one filepath to another
 
-    int get_num_joysticks() const; ///< Get the number of joysticks attached to the system
+    size_t get_num_joysticks() const; ///< Get the number of joysticks attached to the system
 
-    const std::string & get_joystick_name(const int &index) const; ///< Get the name of a given joystick
-    int get_joystick_num_axes(const int &index) const; ///< Get the number of axes for a joystick
-    int get_joystick_num_balls(const int &index) const; ///< Get the number of balls for a joystick
-    int get_joystick_num_hats(const int &index) const; ///< Get the number of hats for a joystick
-    int get_joystick_num_buttons(const int &index) const; ///< Get the number of buttons for a joystick
+    const std::string & get_joystick_name(const size_t &index) const; ///< Get the name of a given joystick
+    int get_joystick_num_axes(const size_t &index) const; ///< Get the number of axes for a joystick
+    int get_joystick_num_balls(const size_t &index) const; ///< Get the number of balls for a joystick
+    int get_joystick_num_hats(const size_t &index) const; ///< Get the number of hats for a joystick
+    int get_joystick_num_buttons(const size_t &index) const; ///< Get the number of buttons for a joystick
 
     void regenerate_joysticks(); ///< Reload all joysticks, flusing *all* SDL events and possibly changing 'which' values for joysticks
